pyproject.toml
[project]
name = "vcache"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "numpy (>=1.26.4,<2.0.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "langchain-openai (>=0.3.14,<0.4.0)",
    "langchain-anthropic (>=0.3.12,<0.4.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "langchain-google-genai (>=2.1.3,<3.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "openai (>=1.75.0,<2.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "hnswlib (>=0.8.0,<0.9.0)",
    "chromadb (>=1.0.5,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "accelerate (>=1.6.0,<2.0.0)",
    "typing-extensions (>=4.13.2,<5.0.0)",
    "torchvision (>=0.22.0,<0.23.0)",
    "statsmodels (>=0.14.4,<0.15.0)",
    "ijson (>=3.4.0,<4.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "datasets (>=3.6.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.benchmarks.dependencies]
matplotlib = "^3.7.0"
ijson = "^3.2.0"
seaborn = "^0.12.0"
scipy = "^1.10.0"
adjustText = "^0.8"
scikit-learn = "^1.2.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.6"
mypy = "^1.15.0"
pre-commit = "^4.2.0"
pytest = "^8.0.0"
python-dotenv = "^1.1.0"


[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false


[tool.mypy]
python_version = "3.10"
packages = ["vcache"]
namespace_packages = true
explicit_package_bases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
